requirements.txt
streamlit==1.45.1
neo4j==5.28.1
cassandra-driver==3.29.2
plotly==6.0.1
numpy==2.2.5
pytest==8.3.5
